+++ coroutine.h
@@ -91,7 +91,7 @@ zend_coroutine_t *async_new_coroutine(zend_async_scope_t *scope);
 void async_coroutine_cleanup(zend_fiber_context *context);
 void async_coroutine_finalize(async_coroutine_t *coroutine);
 void async_coroutine_finalize_from_scheduler(async_coroutine_t *coroutine);
-bool async_coroutine_suspend(const bool from_main);
+bool async_coroutine_suspend(const bool from_main, const bool is_bailout);
 bool async_coroutine_resume(zend_coroutine_t *coroutine, zend_object *error, const bool transfer_error);
 bool async_coroutine_cancel(zend_coroutine_t *zend_coroutine,
 							zend_object *error,